feat: Refactor GrpcUpstreamService to support user-supplied protobufs
This refactors the GrpcUpstreamService to support both gRPC reflection and user-supplied protobuf definitions.

The `proto_content` field has been replaced with a repeated `ProtoDefinition` field, which can be either a `ProtoFile` or a `ProtoDescriptor`.

A `ProtoCollection` message has also been added to allow users to specify a directory to search for protobuf files, with support for recursive searching and regex matching. The `proto_collection` field is repeated to allow for multiple collections.

Unit tests have been added to verify the new functionality.