feat: Use GOTOOLCHAIN for tests
This change updates the Makefile to use the GOTOOLCHAIN environment
variable to specify the Go toolchain for tests. This provides more
control over the build environment and ensures that the correct Go
version is used for testing.

The GOTOOLCHAIN variable is defined at the top of the Makefile for
easy modification.