feat: Refactor command line upstream
Refactors StdioCallDefinition and CommandLineUpstreamService.

- StdioCallDefinition is renamed to CommandLineCallDefinition
- StdioHealthCheck is renamed to CommandLineHealthCheck
- The args field is moved to CommandLineCallDefinition
- A new message `ContainerEnvironment` is created
- A `container_environment` field is added to CommandLineUpstreamService, when set, the command is executed in a docker container
- A `timeout` field is added to CommandLineUpstreamService, which is used to restrict how long the specified command can run

The existing implementation, example and tests are changed to so this latest definition.
In the example and tests, the `datetime` command is used as an example, and CommandLineCallDefinition is configured to return different formats of datetime using different arguments.
The logic for McpStdioConnection is refactored and reused.